+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright 2018-2020 Ian Johnson
+Copyright 2018-2021 Ian Johnson
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of
 this software and associated documentation files (the "Software"), to deal in
