+++ index.html
@@ -2,16 +2,15 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta
       name="viewport"
       content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"
     />
     <link rel="stylesheet" href="./src/index.css" />
     <title>Temperatune</title>
+    <script type="module" src="./src/app.ts"></script>
   </head>
   <body>
     <tt-app></tt-app>
-    <script type="module" src="./src/app.ts"></script>
   </body>
 </html>
