+++ src/index.css
@@ -13,8 +13,6 @@ body {
   margin: 0;
   padding: 0;
 
-  background: #eee;
-
   font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
     Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
 }
