--- src/play-button.ts
@@ -1,67 +1,0 @@
-import { LitElement, html, css } from "lit";
-import { customElement, property } from "lit/decorators.js";
-import { createRef, ref } from "lit/directives/ref.js";
-import playIcon from "./play.svg";
-import micIcon from "./mic.svg";
-
-export class PlayToggleEvent extends Event {
-  constructor(public playing: boolean) {
-    super("toggle", { bubbles: true, composed: true });
-  }
-}
-
-@customElement("tt-play-button")
-export class PlayButton extends LitElement {
-  static override styles = css`
-    :host {
-      display: block;
-    }
-
-    button {
-      height: 100%;
-
-      background: none;
-      border: none;
-      border-radius: 50%;
-
-      transition: background 200ms;
-    }
-
-    button:hover {
-      background: #ccc;
-    }
-
-    img {
-      width: auto;
-      height: 100%;
-    }
-  `;
-
-  @property({ type: Boolean }) playing = false;
-  private _button = createRef<HTMLButtonElement>();
-
-  override render() {
-    return html`
-      <button
-        aria-label=${this.playing ? "Listen for note" : "Play note"}
-        @click=${this._handleClick}
-        ${ref(this._button)}
-      >
-        <img src=${this.playing ? micIcon : playIcon} />
-      </button>
-    `;
-  }
-
-  private _handleClick() {
-    this.playing = !this.playing;
-    this.dispatchEvent(new PlayToggleEvent(this.playing));
-    this._button.value?.animate(
-      [
-        { boxShadow: "none" },
-        { boxShadow: "0 0 20px #1e9be9" },
-        { boxShadow: "none" },
-      ],
-      200
-    );
-  }
-}
